SDF Fusion working pretty well in simulation.

+++ kangaroo/Sdf.h
@@ -8,7 +8,7 @@ namespace Gpu
 struct SDF_t {
     inline __host__ __device__ SDF_t() {}
     inline __host__ __device__ SDF_t(float v) : val(v), w(1) {}
-    inline __host__ __device__ SDF_t(float v, int w) : val(v), w(w) {}
+    inline __host__ __device__ SDF_t(float v, float w) : val(v), w(w) {}
     inline __host__ __device__ operator float() const {
         return val;
     }
